+++ src/test/java/com/github/csalmhof/aoc2021/Puzzle03Test.java
@@ -8,12 +8,12 @@ public class Puzzle03Test extends AbstractPuzzleTest {
   }
 
   @Override
-  public int getExpectedExampleResultPart1() {
+  public long getExpectedExampleResultPart1() {
     return 198;
   }
 
   @Override
-  public int getExpectedExampleResultPart2() {
+  public long getExpectedExampleResultPart2() {
     return 230;
   }
 }
